requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
streamlit==1.49.1
pandas==2.3.2
plotly==6.3.0
requests==2.32.4
fitparse==1.2.0
matplotlib==3.10.5
google-generativeai==0.8.5
